ImageSearchService/src/requirement.txt
fastapi==0.110.0
uvicorn==0.29.0
pillow==10.2.0
numpy==1.26.4
torch==2.3.0
torchvision==0.18.0
elasticsearch==8.12.0
python-multipart==0.0.9